Extract classes which require game installed to new project
