+++ CommandLine/CommandLine.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Worms", "src\Worms\worms.csproj", "{DF42A167-014E-4028-8E3A-12D76A211AB5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Worms.Armageddon.Game", "src\Worms.Armageddon.Game\Worms.Armageddon.Game.csproj", "{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +32,17 @@ Global
 		{DF42A167-014E-4028-8E3A-12D76A211AB5}.Release|x64.Build.0 = Release|Any CPU
 		{DF42A167-014E-4028-8E3A-12D76A211AB5}.Release|x86.ActiveCfg = Release|Any CPU
 		{DF42A167-014E-4028-8E3A-12D76A211AB5}.Release|x86.Build.0 = Release|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Debug|x64.Build.0 = Debug|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Debug|x86.Build.0 = Debug|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Release|x64.ActiveCfg = Release|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Release|x64.Build.0 = Release|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Release|x86.ActiveCfg = Release|Any CPU
+		{D2CEB1E4-3B74-4279-BB35-6FDAE14BB587}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
