+++ CommandLine/src/Worms.Armageddon.Game/IWormsRunner.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace Worms.WormsArmageddon
+namespace Worms.Armageddon.Game
 {
     public interface IWormsRunner
     {
