+++ CommandLine/src/Worms.Armageddon.Game/Linux/WormsLocator.cs
@@ -1,31 +1,30 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
-using System.Runtime.Versioning;
-using Microsoft.Win32;
 
-namespace Worms.WormsArmageddon.Windows
+namespace Worms.Armageddon.Game.Linux
 {
-    [SupportedOSPlatform("windows")]
     internal class WormsLocator : IWormsLocator
     {
         public GameInfo Find()
         {
             const string processName = "WA";
+            var userHomeDirectory = Environment.GetEnvironmentVariable("HOME");
 
-            var location = Registry.GetValue(
-                @"HKEY_CURRENT_USER\SOFTWARE\Team17SoftwareLTD\WormsArmageddon",
-                "Path",
-                null);
-
-            if (location is null || !(location is string rootLocation))
+            if (string.IsNullOrEmpty(userHomeDirectory))
             {
                 return GameInfo.NotInstalled;
             }
 
+            var rootLocation = Path.Combine(userHomeDirectory, "games", "worms");
             var exeLocation = Path.Combine(rootLocation, processName + ".exe");
             var schemesFolder = Path.Combine(rootLocation, "User", "Schemes");
             var gamesFolder = Path.Combine(rootLocation, "User", "Games");
+
+            if (!File.Exists(exeLocation))
+            {
+                return GameInfo.NotInstalled;
+            }
 
             var versionInfo = FileVersionInfo.GetVersionInfo(exeLocation);
             var version = new Version(
