+++ CommandLine/src/Worms.Armageddon.Game/Replays/IReplayLocator.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Worms.WormsArmageddon.Replays
+namespace Worms.Armageddon.Game.Replays
 {
     public interface IReplayLocator
     {
