+++ CommandLine/src/Worms.Armageddon.Game/Replays/IReplayLogGenerator.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace Worms.WormsArmageddon.Replays
+namespace Worms.Armageddon.Game.Replays
 {
     public interface IReplayLogGenerator
     {
