+++ CommandLine/src/Worms.Armageddon.Game/Replays/ReplayLocator.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.IO.Abstractions;
 using System.Linq;
 
-namespace Worms.WormsArmageddon.Replays
+namespace Worms.Armageddon.Game.Replays
 {
     internal class ReplayLocator : IReplayLocator
     {
