+++ CommandLine/src/Worms.Armageddon.Game/Replays/ReplayLogGenerator.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace Worms.WormsArmageddon.Replays
+namespace Worms.Armageddon.Game.Replays
 {
     public class ReplayLogGenerator : IReplayLogGenerator
     {
