+++ CommandLine/src/Worms.Armageddon.Game/Worms.Armageddon.Game.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Autofac" Version="6.1.0" />
+      <PackageReference Include="Microsoft.Win32.Registry" Version="5.0.0" />
+      <PackageReference Include="System.IO.Abstractions" Version="13.2.25" />
+    </ItemGroup>
+
+</Project>
