+++ CommandLine/src/Worms/Commands/Resources/Replays/ProcessReplay.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
-using Worms.WormsArmageddon.Replays;
+using Worms.Armageddon.Game.Replays;
 
 // ReSharper disable MemberCanBePrivate.Global - CLI library uses magic to read members
 // ReSharper disable UnassignedGetOnlyAutoProperty - CLI library uses magic to set members
