+++ CommandLine/src/Worms/Commands/Version.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
+using Worms.Armageddon.Game;
 using Worms.Cli;
-using Worms.WormsArmageddon;
 
 // ReSharper disable UnusedMember.Global - CLI library uses magic to call OnExecuteAsync()
