+++ CommandLine/src/Worms/Container/LinuxCliModule.cs
@@ -1,7 +1,5 @@
 ﻿using Autofac;
 using Worms.Configuration.SecureStorage;
-using Worms.WormsArmageddon;
-using Worms.WormsArmageddon.Linux;
 
 namespace Worms.Container
 {
@@ -10,9 +8,6 @@ namespace Worms.Container
         protected override void Load(ContainerBuilder builder)
         {
             builder.RegisterType<NoOpCredentialStorage>().As<ICredentialStorage>();
-
-            builder.RegisterType<WormsLocator>().As<IWormsLocator>();
-            builder.RegisterType<WormsRunner>().As<IWormsRunner>();
         }
     }
 }
