+++ CommandLine/src/Worms/Container/WindowsCliModule.cs
@@ -1,7 +1,5 @@
 ﻿using Autofac;
 using Worms.Configuration.SecureStorage;
-using Worms.WormsArmageddon;
-using Worms.WormsArmageddon.Windows;
 
 namespace Worms.Container
 {
@@ -10,10 +8,6 @@ namespace Worms.Container
         protected override void Load(ContainerBuilder builder)
         {
             builder.RegisterType<WindowsCredentialStorage>().As<ICredentialStorage>();
-
-            builder.RegisterType<SteamService>().As<ISteamService>();
-            builder.RegisterType<WormsLocator>().As<IWormsLocator>();
-            builder.RegisterType<WormsRunner>().As<IWormsRunner>();
         }
     }
 }
