+++ CommandLine/src/Worms/Resources/Replays/LocalReplayRetriever.cs
@@ -2,8 +2,8 @@
 using System.Collections.Generic;
 using System.IO.Abstractions;
 using System.Linq;
+using Worms.Armageddon.Game.Replays;
 using Worms.Resources.Replays.Text;
-using Worms.WormsArmageddon.Replays;
 
 namespace Worms.Resources.Replays
 {
