+++ CommandLine/src/Worms/Resources/Schemes/LocalSchemeDeleter.cs
@@ -1,5 +1,5 @@
 using System.IO.Abstractions;
-using Worms.WormsArmageddon;
+using Worms.Armageddon.Game;
 
 namespace Worms.Resources.Schemes
 {
